test: Fixup tests to use new package paths for rate limit repository
