--- gravitee-am-repository/gravitee-am-repository-tests/src/test/java/io/gravitee/am/repository/ratelimit/test/config/RateLimitTestConfigurationLoader.java
@@ -1,24 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.am.repository.ratelimit.test.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class RateLimitTestConfigurationLoader {
-    // empty class that will be excluded from the test-jar.
-    // must be created in each repository module to provide relevant spring configuration
-}
